fix terrifying recursion
